Use prebuilt noconflict version of ace
It contains a build that doesn't use requirejs. Instead, it uses a
custom amd shim, such that within ace `require` is replaced by
`ace.require`. This is known as the "noconflict" version.

Loading the normal version of require doesn't work for us because:

1) We use almond to optimize requirejs away in the bundled build.
2) The normal version of ace expects normal requirejs
3) Because of 2), ace uses require.toURL to load asset.
4) toURL isn't supported by almond.

The upshot that we need to load ace as though it is a global script.

Ace modes, themes, and extensions are normally loaded on the fly. In our
case, we need to load them manually if we are to use them, so they are
accessible using the `ace.require` shim.